Huge refactor
- move fixtures and templates into p2panda-rs crate in their own module
- leave only Panda dummy node/client and test data generation in p2panda-tests
