+++ p2panda-rs/Cargo.toml
@@ -39,6 +39,8 @@ serde_cbor = "0.11.2"
 serde_repr = "0.1.7"
 thiserror = "1.0.29"
 yamf-hash = "0.1.2"
+rstest = "0.11.0"
+rstest_reuse = "0.1.3"
 
 # Optional dependencies (behind feature flags)
 sqlx = { version = "0.5.9", features = ["runtime-async-std-rustls", "all-databases"], optional = true }
@@ -54,4 +56,3 @@ wasm-bindgen = "0.2.78"
 
 [dev-dependencies]
 wasm-bindgen-test = "0.3.28"
-rstest = "0.11.0"
