+++ p2panda-rs/src/tests/mod.rs
@@ -1,0 +1,4 @@
+
+pub mod templates;
+pub mod fixtures;
+pub mod utils;
