+++ p2panda-tests/src/lib.rs
@@ -1,25 +1,14 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
-mod fixtures;
-pub mod generate_test_data;
-mod templates;
-#[cfg(test)]
-mod tests;
 pub mod utils;
 
-// This must be imported here at the root of the crate.
-#[allow(unused_imports)]
-use rstest_reuse;
-
-use crate::utils::NextEntryArgs;
-
-use bamboo_rs_core::entry::is_lipmaa_required;
 use std::collections::HashMap;
-use std::convert::{TryFrom, TryInto};
+use std::convert::TryFrom;
 
 use p2panda_rs::entry::{decode_entry, sign_and_encode, Entry, EntrySigned, LogId, SeqNum};
-use p2panda_rs::hash::Hash;
 use p2panda_rs::identity::{Author, KeyPair};
-use p2panda_rs::message::{Message, MessageEncoded, MessageFields, MessageValue};
+use p2panda_rs::message::{Message, MessageEncoded};
+
+use crate::utils::{calculate_entry_args, NextEntryArgs};
 
 // const META_SCHEMA: &str  = "004069db5208a271c53de8a1b6220e6a4d7fcccd89e6c0c7e75c833e34dc68d932624f2ccf27513f42fb7d0e4390a99b225bad41ba14a6297537246dbe4e6ce150e8";
 
@@ -55,21 +44,6 @@ impl Panda {
         hex::encode(self.key_pair.public_key())
     }
 
-    pub fn some_hash(str: &str) -> Option<Hash> {
-        let hash = Hash::new(str);
-        Some(hash.unwrap())
-    }
-
-    pub fn build_message_fields(fields: Vec<(&str, &str)>) -> MessageFields {
-        let mut message_fields = MessageFields::new();
-        for (key, value) in fields.iter() {
-            message_fields
-                .add(key, MessageValue::Text(value.to_string()))
-                .unwrap();
-        }
-        message_fields
-    }
-
     pub fn name(&self) -> String {
         self.name.to_owned()
     }
@@ -77,7 +51,7 @@ impl Panda {
     /// Calculate the next entry arguments for this log (log_id, seq_num, backlink, skiplink)
     pub fn next_entry_args(&self, log_id: usize) -> NextEntryArgs {
         let schema_entries = self.logs.get(&log_id).unwrap();
-        Panda::calculate_entry_args(log_id, schema_entries.to_owned())
+        calculate_entry_args(log_id, schema_entries.to_owned())
     }
 
     /// Calculate the next entry arguments *at a certain point* in this log. This is helpful
@@ -89,7 +63,7 @@ impl Panda {
         seq_num: &SeqNum,
     ) -> NextEntryArgs {
         let schema_entries = self.logs.get(&log_id).unwrap();
-        Panda::calculate_entry_args(
+        calculate_entry_args(
             log_id,
             schema_entries[..seq_num.as_i64() as usize - 1].to_owned(),
         )
@@ -113,12 +87,15 @@ impl Panda {
         let next_entry_args = self.next_entry_args(log_id);
 
         // Construct entry from message and entry args then sign and encode it
-        let entry = Panda::entry(
-            &message,
+        let entry = Entry::new(
+            &LogId::default(),
+            Some(&message),
+            next_entry_args.entryHashBacklink.as_ref(),
+            next_entry_args.entryHashBacklink.as_ref(),
             &next_entry_args.seqNum,
-            next_entry_args.entryHashBacklink,
-            next_entry_args.entryHashSkiplink,
-        );
+        )
+        .unwrap();
+
         let entry_encoded = sign_and_encode(&entry, &self.key_pair).unwrap();
 
         // Encode message
@@ -143,110 +120,5 @@ impl Panda {
     ) -> (EntrySigned, MessageEncoded) {
         let log_id = self.schema.get(schema).unwrap();
         self.logs.get(log_id).unwrap()[seq_num - 1].clone()
-    }
-}
-
-// These are all static helper methods mainly used within the testing crate for easily creating common values.
-impl Panda {
-    pub fn entry(
-        message: &Message,
-        seq_num: &SeqNum,
-        backlink: Option<Hash>,
-        skiplink: Option<Hash>,
-    ) -> Entry {
-        Entry::new(
-            &LogId::default(),
-            Some(message),
-            skiplink.as_ref(),
-            backlink.as_ref(),
-            seq_num,
-        )
-        .unwrap()
-    }
-
-    pub fn message(
-        schema: &str,
-        instance_id: Option<Hash>,
-        fields: Option<Vec<(&str, &str)>>,
-    ) -> Message {
-        match fields {
-            // It's a CREATE message
-            Some(fields) if instance_id.is_none() => Panda::create_message(schema, fields),
-            // It's an UPDATE message
-            Some(fields) => Panda::update_message(schema, instance_id.unwrap(), fields),
-            // It's a DELETE message
-            None => Panda::delete_message(schema, instance_id.unwrap()),
-        }
-    }
-
-    fn calculate_entry_args(
-        log_id: usize,
-        schema_entries: Vec<(EntrySigned, MessageEncoded)>,
-    ) -> NextEntryArgs {
-        if schema_entries.len() == 0 {
-            NextEntryArgs {
-                entryHashBacklink: None,
-                entryHashSkiplink: None,
-                seqNum: Panda::seq_num(1),
-                logId: LogId::new(log_id.try_into().unwrap()),
-            }
-        } else {
-            // Get last entry in log
-            let (entry_encoded, _) = schema_entries.get(schema_entries.len() - 1).unwrap();
-            let decoded_last_entry = decode_entry(&entry_encoded, None).unwrap();
-
-            // Get the hash (which is the backlink we need)
-            let backlink = Some(entry_encoded.hash());
-
-            // Get the next sequence number
-            let next_seq_num = decoded_last_entry.seq_num().to_owned().next().unwrap();
-
-            // And then the skiplink
-            let skiplink_seq_num = next_seq_num.skiplink_seq_num().unwrap().as_i64();
-
-            // And finally the log id
-            let log_id = decoded_last_entry.log_id();
-
-            // Check if skiplink is required and return hash if so
-            let skiplink = if is_lipmaa_required(next_seq_num.as_i64() as u64) {
-                let (skiplink_entry, _) = schema_entries.get(skiplink_seq_num as usize).unwrap();
-                Some(skiplink_entry.hash())
-            } else {
-                None
-            };
-
-            NextEntryArgs {
-                entryHashBacklink: backlink,
-                entryHashSkiplink: skiplink,
-                seqNum: next_seq_num,
-                logId: log_id.to_owned(),
-            }
-        }
-    }
-
-    pub fn create_message(schema: &str, fields: Vec<(&str, &str)>) -> Message {
-        let fields = Panda::build_message_fields(fields);
-        Message::new_create(Hash::new(schema).unwrap(), fields).unwrap()
-    }
-
-    pub fn update_message(schema: &str, instance_id: Hash, fields: Vec<(&str, &str)>) -> Message {
-        let fields = Panda::build_message_fields(fields);
-        Message::new_update(Hash::new(schema).unwrap(), instance_id, fields).unwrap()
-    }
-
-    pub fn delete_message(schema: &str, instance_id: Hash) -> Message {
-        Message::new_delete(Hash::new(schema).unwrap(), instance_id).unwrap()
-    }
-
-    pub fn seq_num(seq_num: i64) -> SeqNum {
-        SeqNum::new(seq_num).unwrap()
-    }
-
-    pub fn keypair_from_private(private_key: &str) -> KeyPair {
-        KeyPair::from_private_key_str(private_key).unwrap()
-    }
-
-    pub fn keypair() -> KeyPair {
-        KeyPair::new()
     }
 }
